mako: touch: preserving touch data coherency
Synaptics firmware maintains internal cache to store a scan of
finger data so that the host reads finger data from this cache.
The firmware updates the cache when the host reads it nonsequentially.

The current driver reads only present finger data, which can cause
unexpected nonsequential access to cache and the cache can be updated
before the host reads the data, resulting in losing finger data coherency.

Modified the driver to access data sequentially even if there are
not present fingers in the middle so that the finger data coherency
can be preserved.

Change-Id: I4b975c2ab2445078f2a81c3dd4ce484002d27933
Signed-off-by: Iliyan Malchev <malchev@google.com>
Signed-off-by: klozz <xxx.reptar.rawrr.xxx@gmail.com>