input: touch: fix sy3200 debug print corruption
Properly handle the debug string format printing as the previous call
to sprintf would always corrupt the stack since it appends a NULL
termination.  There is no need for a local buffer and a second strcpy.
This would lead to kernel panics randomly with backtrace due to the
'dd' pointer being corrupted.

Change-Id: I46e7407099e0c2e6d9d230675b0049fa72726974
Signed-off-by: Jason Hrycay <jason.hrycay@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/509300
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Christopher Fries <qcf001@motorola.com>
Signed-off-by: klozz <xxx.reptar.rawrr.xxx@gmail.com>